Create an FxABrowserIDClient trait
