+++ components/fxa-client/src/http_client/browser_id/hawk_request.rs
